requirements.txt
langchain-huggingface
huggingface_hub
transformers
accelerate
bitsandbytes
streamlit
langchain
groq
langchain-groq
langchain_community

faiss-cpu

langchain_google_genai
python-dotenv
google-cloud-aiplatform>=1.38
chardet
unstructured
unstructured[pdf]


git+https://github.com/openai/whisper.git